% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlagFigAndelerGrVar.R
\name{FigAndelerGrVar}
\alias{FigAndelerGrVar}
\title{Søylediagram med andeler for hver grupperingsenhet (sykehus, RHF, ...)}
\usage{
FigAndelerGrVar(RegData, valgtVar, datoFra = "2012-04-01",
  datoTil = "2050-12-31", enhetsUtvalg = 0, minald = 0, maxald = 130,
  erMann = "", diagnose = "", innl4t = "", NIHSSinn = "",
  hentData = 0, preprosess = 1, reshID = 0, outfile = "")
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres
BehSlagenhet: Behandlet i slagenhet
InnlInnen4eSymptom: Innlagt innen 4t etter symptomdebut
InnlSlagenh: Innlagt direkte i slagehet
LipidI63u80: Utskrevet med lipidsenkning
SvelgtestUtfort: Svelgfunksjon vurdert
TidInnTrombolyse30min:  Trombolyse innen 30 min. etter innleggelse
TrombolyseI63: Hjerneinfarktpasienter som har fått trombolyse
UtBT: Blodtrykksmedikament ved utskriving
UtAntitrombotiskI63: Utskrevet med antitrombotisk behandling. Innleggelser etter 31.12.2013.
UtAntikoagI63atrie: Utskrevet med antikoagulasjon (hjerneinfarktpasienter med atrieflimmer)}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{enhetsUtvalg}{Gjør gruppeutvalg for
0: Hele landet
     7: Egen region}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{diagnose}{1: Infarkt(I61), 
2: Blødning(I63), 
3: definert(I64), 
standard: alle (ikke spesifisert)}

\item{innl4t}{innlagt innen 4 timer. 
0: nei
1: ja, standard:'' (alt annet)}

\item{NIHSSinn}{NIHSS ved innkomst. Registreringer hvor NIHSS ikke er utførtblir tatt bort. 
Kategorier: 0-5,6-10,11-15,...}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{preprosess}{Preprosesser data
FALSE: Nei
TRUE: Ja (Standard)}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}
}
\value{
Søylediagram med andeler av valgt variabel for hvert sykehus
}
\description{
Funksjon som genererer en figur med andeler av en variabel for en grupperingsvariabelen sykehus.
Funksjonen er delvis skrevet for å kunne brukes til andre grupperingsvariable enn sykehus
}
\details{
Detajer: Her bør man liste opp hvilke variable funksjonen benytter...
}

