% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlagFigGjsnTid.R
\name{FigGjsnTid}
\alias{FigGjsnTid}
\title{Utvikling over tid for gjennomsnitt/median av valgt variabel}
\usage{
FigGjsnTid(RegData, valgtVar, datoFra = "2013-01-01",
  datoTil = "3000-12-31", minald = 0, maxald = 130, erMann = "",
  diagnose = "", innl4t = "", NIHSSinn = "", reshID, outfile = "",
  enhetsUtvalg = 1, valgtMaal = "", preprosess = 1, hentData = 0)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres
Alder: Alder
AntDagerInnl: liggetid
NIHSSinnkomst: NIHSS ved innkomst
NIHSSpreTrombolyse: NIHSS før trombolyse
NIHSSetterTrombolyse: NIHSS etter trombolyse
TidInnleggTrombolyse: Tid fra innleggelse til trombolyse
TidSymptInnlegg: Tid fra symptom til innleggelse
TidSymptTrombolyse: Tid fra symptom til trombolyse}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{diagnose}{1: Infarkt(I61), 
2: Blødning(I63), 
3: definert(I64), 
standard: alle (ikke spesifisert)}

\item{innl4t}{innlagt innen 4 timer. 
0: nei
1: ja, standard:'' (alt annet)}

\item{NIHSSinn}{NIHSS ved innkomst. Registreringer hvor NIHSS ikke er utførtblir tatt bort. 
Kategorier: 0-5,6-10,11-15,...}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}

\item{enhetsUtvalg}{Gjør gruppeutvalg med eller uten sammenlikning for
 0: Hele landet
 1: Egen enhet mot resten av landet (Standard)
 2: Egen enhet
6: Egen enhet mot egen region 
7: Egen region 
8: Egen region mot resten}

\item{valgtMaal}{'Gjsn': gir middelverdi (standard)
'Med': gir median}

\item{preprosess}{Preprosesser data
FALSE: Nei
TRUE: Ja (Standard)}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}
}
\value{
Linjediagram som viser utvikling over tid for valgt variabel
}
\description{
Figuren viser gjennomsnitt/median per år med konfidensintervall for valgt variabel.
I bakgrunn vises konfidensintervall for resten av landet.
}
\details{
Detajer: Her bør man liste opp hvilke variable funksjonen benytter...
}

