% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlagFigAndelerPrePost.R
\name{FigPrePost}
\alias{FigPrePost}
\title{Søylediagram som fordeling av valgt variabel, målt både før og etter hjerneslaget}
\usage{
FigPrePost(RegData, valgtVar, datoFra = "2012-04-01",
  datoTil = "2050-12-31", minald = 0, maxald = 130, erMann = "",
  diagnose = "", innl4t = "", NIHSSinn = "", libkat, outfile = "",
  reshID, enhetsUtvalg = 1, preprosess = 1, hentData = 0)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres
Boligforh: Boligforhold ved innleggelse og 3 mnd etter slaget
Bosituasjon: Bosituasjon ved innleggelse og 3 mnd etter slaget
Toalett: Toalettbesok før innleggelse og 3 mnd etter slaget 
Forflytning: Evne til forflytning før innleggelse og 3 mnd etter slaget
Paakledning: Evne til egen påkledning før innleggelse og 3 mnd etter slaget
Bilkjoring: Bilkjøring før innleggelse og 3 mnd etter slaget
Yrkesaktiv: Yrkesdeltagelse før innleggelse og 3 mnd etter slaget
Roykestatus: Røykestatus før innleggelse og 3 mnd etter slaget
MRS: "modified Rankin scale" ved slaget og 3 mnd etter
NIHSSTrombolyse: NIHSS før og etter trombolyse
NIHSSTrombektomi: NIHSS før og etter trombektomi}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{diagnose}{1: Infarkt(I61), 
2: Blødning(I63), 
3: definert(I64), 
standard: alle (ikke spesifisert)}

\item{innl4t}{innlagt innen 4 timer. 
0: nei
1: ja, standard:'' (alt annet)}

\item{NIHSSinn}{NIHSS ved innkomst. Registreringer hvor NIHSS ikke er utførtblir tatt bort. 
Kategorier: 0-5,6-10,11-15,...}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{enhetsUtvalg}{Gjør gruppeutvalg med eller uten sammenlikning for
 0: Hele landet
 1: Egen enhet mot resten av landet (Standard)
 2: Egen enhet
6: Egen enhet mot egen region 
7: Egen region 
8: Egen region mot resten}

\item{preprosess}{Preprosesser data
FALSE: Nei
TRUE: Ja (Standard)}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}
}
\value{
Søylediagram som fordeling av valgt variabel, målt både før og etter hjerneslaget
}
\description{
Funksjon som genererer en figur med som viser endring i en variabels fordeling fra før til etter hjerneslaget.
}
\details{
Detajer: Her bør man liste opp hvilke variable funksjonen benytter...
}

