% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlagFigAndeler.R
\name{SlagFigAndeler}
\alias{SlagFigAndeler}
\title{Søylediagram som viser andeler av ulike variabler}
\usage{
SlagFigAndeler(RegData, valgtVar, datoFra = "2012-04-01",
  datoTil = "2050-12-31", minald = 0, maxald = 130, erMann = "",
  diagnose = "", innl4t = "", NIHSSinn = "", outfile = "",
  preprosess = 1, hentData = 0, reshID, enhetsUtvalg = 1)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres. Se \strong{Details} for oversikt.}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{diagnose}{1: Infarkt(I61), 
2: Blødning(I63), 
3: definert(I64), 
standard: alle (ikke spesifisert)}

\item{innl4t}{innlagt innen 4 timer. 
0: nei
1: ja, standard:'' (alt annet)}

\item{NIHSSinn}{NIHSS ved innkomst. Registreringer hvor NIHSS ikke er utførtblir tatt bort. 
Kategorier: 0-5,6-10,11-15,...}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}

\item{preprosess}{Preprosesser data
FALSE: Nei
TRUE: Ja (Standard)}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{enhetsUtvalg}{Gjør gruppeutvalg med eller uten sammenlikning for
 0: Hele landet
 1: Egen enhet mot resten av landet (Standard)
 2: Egen enhet
6: Egen enhet mot egen region 
7: Egen region 
8: Egen region mot resten}
}
\value{
Søylediagram (fordeling) av valgt variabel. De enkelte verdiene kan også sendes med.
}
\description{
Denne funksjonen lager et søylediagram som viser andeler (fordeling) av valgt variabel
filtrert på de utvalg som er gjort.
}
\details{
Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item Alder: Aldersfordeling, 10-årige grupper 
    \item AntDagerInnl: Liggetid 
    \item AvdForstInnlagtHvilken: Hvilken avdeling ble pasienten først innlagt?
    \item AvdUtskrFraHvilken: Hvilken avdeling ble pasienten utskrevet fra? 
    \item BevissthetsgradInnleggelse: Bevissthetsgrad ved innleggelsen 
    \item BildediagnostikkEkstrakranKar: Bildediagnostikk av ekstrakranielle kar 
    \item BildediagnostikkHjerne: Bildediagnostikk av hjerneslaget
    \item BildediagnostikkHjerte: Bildediagnostikk av hjertet 
    \item BildediagnostikkIntraraniell: Bildediagnostikk av intrakranielle kar
    \item Boligforhold3mnd: Boligforhold ved oppfølging 
    \item BoligforholdPre: Boligforhold ved innleggelse 
	\item FokaleUtf: Fokale utfall
	\item FokaleUtfAndre: Andre fokale utfall
    \item MRS3mnd: Rankinscale ved oppfølging
    \item MRSPre: Rankinscale ved innleggelse 
	\item NIHSSendrTrombektomi: Endring i NIHSS fra før trombektomi til 24t etter
	\item NIHSSendrTrombolyse: Endring i NIHSS fra før trombolyse til 24t etter
    \item NIHSSetterTrombektomi: NIHSS 24t etter trombektomi
    \item NIHSSetterTrombolyse: NIHSS 24t etter trombolyse
    \item NIHSSinnkomst: NIHSS ved innkomst
    \item NIHSSpreTrombektomi: NIHSS før trombektomi
    \item NIHSSpreTrombolyse: NIHSS før trombolyse
    \item RegistreringHjerterytme: Registrering av hjerterytme 
    \item Royker3mnd: Røykestatus ved oppfølging 
    \item RoykerPre: Røykestatus ved innleggelse 
    \item Sivilstatus3mnd: Sivilstatus ved oppfølging
    \item SivilstatusPre: Sivilstatus ved innleggelse 
    \item Slagdiagnose: Slagdiagnose 
    \item TidInnleggTrombolyse: Antall timer fra innleggelse til trombolyse
    \item TidSymptInnlegg: Tid fra symptomdebut til innleggelse', '(kun de som ikke våknet med symptom)
    \item TidSymptTrombolyse: Tid fra symptomdebut til trombolyse,', '(kun de som ikke våknet med symptom)
    \item Tilfredshet: Er du like fornøyd med tilværelsen som før hjerneslaget?
    \item Transportmetode: Transport til sykehus etter varsling av AMK 
    \item UtskrTil. Hva pasientene ble utskrevet til
   }
}

