% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlagPreprosesser.R
\name{SlagPreprosess}
\alias{SlagPreprosess}
\title{Preprosesser data fra Hjerneslag}
\usage{
SlagPreprosess(RegData = RegData, reshID = reshID)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}
}
\value{
Data En list med det filtrerte datasettet og sykehusnavnet som tilsvarer reshID
}
\description{
Denne funksjonen definerer variabler og fjerner ikke-ferdigstilte registreringer
}

