% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlagRegDataSQL.R
\name{SlagRegDataSQL}
\alias{SlagRegDataSQL}
\title{Henter data registrert for Hjerneslag}
\usage{
SlagRegDataSQL(datoFra = "2013-01-01", datoTil = "2099-01-01")
}
\arguments{
\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}
}
\value{
Henter dataramma RegData for Hjerneslag
}
\description{
Henter data for Hjerneslagregisteret fra "staging"
}

