% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlagUtvalg.R
\name{SlagUtvalg}
\alias{SlagUtvalg}
\title{Funksjon som gjør utvalg av dataene, returnerer det filtrerte datasettet og utvalgsteksten.}
\usage{
SlagUtvalg(RegData, datoFra, datoTil, minald, maxald, erMann, diagnose, innl4t,
  NIHSSinn, fargepalett = "BlaaRapp")
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{diagnose}{1: Infarkt(I61), 
2: Blødning(I63), 
3: definert(I64), 
standard: alle (ikke spesifisert)}

\item{innl4t}{innlagt innen 4 timer. 
0: nei
1: ja, standard:'' (alt annet)}

\item{NIHSSinn}{NIHSS ved innkomst. Registreringer hvor NIHSS ikke er utførtblir tatt bort. 
Kategorier: 0-5,6-10,11-15,...}

\item{fargepalett}{Hvilken fargepalett skal brukes i figurer (Standard: BlaaRapp)}
}
\value{
UtData En liste bestående av det filtrerte datasettet, utvalgstekst for figur og tekststreng som angir fargepalett
}
\description{
Funksjon som gjør utvalg av dataene, returnerer det filtrerte datasettet og utvalgsteksten.
}

