% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlagFigAndelerKvalInd.R, R/SlagFigAndelerKvalIndOKmenGML.R
\name{FigAndelerKvalInd}
\alias{FigAndelerKvalInd}
\title{Søylediagram som viser andeler av kvalitetsindikatorer}
\usage{
FigAndelerKvalInd(RegData, datoFra = "2012-04-01", datoTil = "2050-12-31",
  erMann = "", NIHSSinn = "", outfile = "", enhetsUtvalg = 1,
  preprosess = 1, hentData = 0, reshID)

FigAndelerKvalInd(RegData, datoFra = "2012-04-01", datoTil = "2050-12-31",
  erMann = "", NIHSSinn = "", outfile = "", enhetsUtvalg = 1,
  preprosess = 1, hentData = 0, reshID)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{erMann}{Kjønn, standard: alt annet enn 0/1 gir begge kjønn
0: Kvinner
1: Menn}

\item{NIHSSinn}{NIHSS ved innkomst. Registreringer hvor NIHSS ikke er utførtblir tatt bort. 
Kategorier: 0-5,6-10,11-15,...}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}

\item{enhetsUtvalg}{Gjør gruppeutvalg med eller uten sammenlikning for
 0: Hele landet
 1: Egen enhet mot resten av landet (Standard)
 2: Egen enhet
6: Egen enhet mot egen region 
7: Egen region 
8: Egen region mot resten}

\item{preprosess}{Preprosesser data
FALSE: Nei
TRUE: Ja (Standard)}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{valgtVar}{Hvilken variabel som skal visualiseres
            BehSlagenhet
            InnlSlagenh
            InnlInnen4eSymptom
            LipidI63
            OppfolgUtf (mangler)
            SvelgtestUtfort
        TidInnTromb30min (mangler)
            TrombolyseI63
            UtAntitrombotiskI63
            UtAntikoagI63atrie
            UtBT

Mye hardkoding/spesialtilpasning. Ikke så egnet til gjenbruk.

Detajer: Her bør man liste opp hvilke variable funksjonen benytter.}
}
\value{
Figur som viser andeler av kvalitetsindikatorer.

Figur som viser andeler av kvalitetsindikatorer.
}
\description{
Funksjonen lager et søylediagram som viser kvalitetsindikatorer fra hjerneslagregisteret
Mye hardkoding/spesialtilpasning. Ikke så egnet til gjenbruk.

Funksjonen lager et søylediagram som viser kvalitetsindikatorer fra hjerneslagregisteret
}
\details{
Detaljer: Her bør man liste opp hvilke variable funksjonen benytter.
}

